refactor(traceback_capture): move TracebackException payload building to submodule
Isolate TracebackException-to-schema conversion logic, including chained exceptions and exception-group extraction, into a new submodule traceback_payload.rs. This cleans up traceback_capture.rs by moving detailed payload construction into its own file, improving code organization and maintainability.
